enhance vmem allocator and slab cache
this patch introduces several new features.
1. implementation of a kvmalloc like interface (kvmem_alloc/kvmem_zalloc)

2. fixes vmem_alloc implementation which can lead to slow allocations
since the first attempt with kmalloc does not make use of the noretry
flag but tells the linux kernel to retry several times before it fails.

3. implements KMC_KVMEM flag for kmem_cache_create which makes use of
a kvmalloc allocator implementation which may increase performance if
the allocator is able to allocate physical memory. otherwise it will
still make use of virtual memory.

4. make use of _GFP_HIGHMEM flag in vmem_alloc , which is required for
some 32 bit systems to make use of full available memory.

5. implement KM_ONCE flag which is behaving almost like KM_SLEEP but
does not hang in a deadloop if no memory is available. it should be used
if you want a GFP_KERNEL like memory allocation, but not that lazy like
KM_NOSLEEP which makes use of GFP_ATOMIC which is mainly made to be used
in irqcontexts. at some locations in zfs like compression algorithms
for instance GFP_ATOMIC/KM_NOSLEEP should not  be used since it might
always fail if blocks are greater than 64-128kb (workspaces). but we also
want no deadlock situation if memory is under pressure. so we need a way
that it behaves like straight GFP_KERNEL. this flag does allow this behaviour.

6. change vmem_alloc to kvmem_alloc on some usefull locations where
memory is initialized.

7. make use of KMC_KVMEM for spl_zlib_workspace_cache, ddt_cache,
zio_buf and zio_data_buf

Signed-off-by: Sebastian Gottschall <s.gottschall@dd-wrt.com>