clean up some mess
rename fields and structures to sane names. some of the older names had
a historical context which is not true anymore.
in addition remove clearing of preallocated buffers on release,
according to the zstd documentation, zstd is able to reuse preinialized
contexts for improving performance. we do not need to clear them anyway

Signed-off-by: Sebastian Gottschall <s.gottschall@dd-wrt.com>
Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>