Add zstd support to zfs
ZStandard is a modern, high performance, general compression algorithm
originally developed by Yann Collet (the author of LZ4). It aims to
provide similar or better compression levels to GZIP, but with much
better performance. ZStandard provides a large selection of compression
levels to allow a storage administrator to select the preferred
performance/compression trade-off.

This PR adds the new compression types: zstd and zstd-fast, which is
basically a faster "negative" level of zstd.

Available compression levels for zstd are zstd-1 through zstd-19. In
addition, faster levels supported by zstd-fast are 1-10, 20-100 (in
increments of 10), 500 and 1000. As zstd-fast is basically a "negative"
level of zstd, it gets faster with every level in stark contrast with
"normal" zstd which gets slower with every level.

Rather than the approach used when gzip was added, of using a separate
compression type for each different level, this PR added a new hidden
property, compression_level. All of the zstd levels use the single entry
in the zio_compress enum, since all levels use the same decompression
function.

However, in some cases, it was necessary to know what level a block was
compressed within other parts of the code, so the compression level is
stored on disk inline with the data, after the compression header
(uncompressed size). The level is plumbed through arc_buf_hdr_t,
objset_t, and zio_prop_t. There may still be additional places where
this needs to be added.

This new property is hidden from the user, and the user simply does zfs
set compress=zstd-10 dataset and zfs_ioctl.c splits this single value
into the two values to be stored in the dataset properties. This logic
has been extended to Channel Programs as well.

Closes: #6247
Co-authored-by: Allan Jude <allanjude@freebsd.org>
Co-authored-by: Sebastian Gottschall <s.gottschall@dd-wrt.com>
Co-authored-by: Michael Niewöhner <foss@mniewoehner.de>
Signed-off-by: Allan Jude <allanjude@freebsd.org>
Signed-off-by: Sebastian Gottschall <s.gottschall@dd-wrt.com>
Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>