Replace the term cookie
In the original version of this code, the 'cookie' was the compression
header, that is now zstd_header_t. Over the course of its evolution
the variable 'cookie' has become the signed integer level passed to
ZSTD for compression, so rename it to what it actually represents.

While here, fix the comment around zstd_enum_to_cookie(),
It incorrectly stated:
"Convert ZSTD level to ZFS internal, stored level"

when it does the opposite, converting the internal ZFS enum to
the signed integer used by ZSTD.

Sponsored-By: The FreeBSD Foundation
Signed-off-by: Allan Jude <allanjude@freebsd.org>