Remove "All Rights Reserved" from copyright header.
As previously explained clearly by the FreeBSD foundation, All Rights Reserved is not needed anymore, due to changes in International law.
Copyright headers continue to function just as a reference of copyright, not as a tool of copyright.
Hence there is no legal implication preventing us from cleaning this up and removing the (now redundant) references to "All Rights Reserved" (which has become the legal default anyway)

See also:
https://svnweb.freebsd.org/base?view=revision&revision=333391

Signed-off-by: Kjeld Schouten-Lebbing <kjeld@schouten-lebbing.nl>