Add a new test to confirm that the ZSTD feature flag is activated
When the compression property on a dataset is set to zstd, the feature
flag is not changed from 'enabled' to 'active' until the first block
is born.

If the pool is exported after `zfs set` but before a block is written
then imported on a system with an older version of ZFS that does not
understand ZSTD, it will cause the userland utilities to panic.

Signed-off-by: Allan Jude <allanjude@freebsd.org>